+++ data/bob.txt
@@ -1,3 +1,3 @@
 todo | 1 | read
-todo | 0 | read
-todo | 1 | hih
+todo | 0 | hih
+todo | 0 | g out
