+++ src/main/java/bob/TaskList.java
@@ -1,4 +1,7 @@
+package bob;
 import java.util.ArrayList;
+import bob.tasks.*;
+import bob.exceptions.*;
 
 /**
  * Contains list of tasks, and operations that alter it.
