+++ src/main/java/bob/commands/AddCommand.java
@@ -1,4 +1,6 @@
+package bob.commands;
 import java.io.IOException;
+import bob.*;
 
 public class AddCommand extends Command {
     public AddCommand(String input) {
