+++ src/main/java/bob/exceptions/BobException.java
@@ -1,0 +1,6 @@
+package bob.exceptions;
+
+public class BobException extends Exception {
+    public String message = "Error.";
+}
+
