+++ src/main/java/bob/exceptions/MissingDeadlineException.java
@@ -1,0 +1,5 @@
+package bob.exceptions;
+
+public class MissingDeadlineException extends BobException {
+    public String message = "A due date must be given for deadlines!";
+}
