+++ src/main/java/bob/exceptions/MissingEventDatesException.java
@@ -1,0 +1,5 @@
+package bob.exceptions;
+
+public class MissingEventDatesException extends BobException {
+    public String message = "Start and end times must be provided for events!";
+}
