+++ src/main/java/bob/exceptions/MissingIndexException.java
@@ -1,0 +1,5 @@
+package bob.exceptions;
+
+public class MissingIndexException extends BobException {
+    public String message = "Task index must be provided!";
+}
