+++ src/main/java/bob/exceptions/MissingTaskException.java
@@ -1,0 +1,5 @@
+package bob.exceptions;
+
+public class MissingTaskException extends BobException {
+    public String message = "Task description is missing!";
+}
