+++ src/main/java/bob/exceptions/WrongInputException.java
@@ -1,0 +1,5 @@
+package bob.exceptions;
+
+public class WrongInputException extends BobException {
+    public String message = "Oh no, I don't understand that...";
+}
