+++ src/main/java/bob/tasks/Deadline.java
@@ -1,5 +1,7 @@
+package bob.tasks;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import bob.exceptions.*;
 
 public class Deadline extends Task {
     private LocalDate due;
