+++ src/main/java/bob/tasks/Task.java
@@ -1,6 +1,4 @@
-enum TaskType {
-    todo, event, deadline
-}
+package bob.tasks;
 
 public class Task {
     protected String name;
