+++ src/main/java/bob/tasks/TaskType.java
@@ -1,0 +1,5 @@
+package bob.tasks;
+
+public enum TaskType {
+    todo, event, deadline
+}
